fix(builder): Re-export anstyle for easy access
