+++ clap_builder/src/builder/mod.rs
@@ -23,6 +23,11 @@ mod debug_asserts;
 #[cfg(test)]
 mod tests;
 
+#[cfg(feature = "unstable-styles")]
+pub mod styling;
+#[cfg(not(feature = "unstable-styles"))]
+pub(crate) mod styling;
+
 pub use self::str::Str;
 pub use action::ArgAction;
 pub use arg::Arg;
@@ -36,7 +41,7 @@ pub use resettable::IntoResettable;
 pub use resettable::Resettable;
 pub use styled_str::StyledStr;
 #[cfg(feature = "unstable-styles")]
-pub use styled_str::Styles;
+pub use styling::Styles;
 pub use value_hint::ValueHint;
 pub use value_parser::_AutoValueParser;
 pub use value_parser::via_prelude;
@@ -63,4 +68,4 @@ pub(crate) use action::CountType;
 pub(crate) use arg_settings::{ArgFlags, ArgSettings};
 pub(crate) use command::AppTag;
 #[cfg(not(feature = "unstable-styles"))]
-pub(crate) use styled_str::Styles;
+pub(crate) use styling::Styles;
